back/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart==0.0.9
pillow>=9.3.0,<10.0.0
torch==2.1.0
diffusers==0.23.1
transformers==4.35.2
accelerate==0.24.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx>=0.24.1,<0.25.0
httpcore>=0.15.0,<0.18.0
python-dotenv==1.0.0
pydantic>=2.6.1
stripe==2.71.0
redis>=5.0.0
web3>=6.0.0
eth-account>=0.8.0
numpy>=1.24.0